Create new algorithm module in infra and move snowflake, uuid to this module
